Update even or odd.py

+++ Beginner/even or odd.py
@@ -1,6 +1,6 @@
 b= input("enter the input:")
 if b.isdigit():
-    if (int(b)%2==0):
+    if (float(b)%2==0):
         print("The given number is even")
     else:
         print("The given interger is odd")
